requirements.txt
git+https://github.com/Seeed-Studio/grove.py
APScheduler==3.9.1.post1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
grovepi==1.0.4
idna==3.4
mfrc522==0.0.7
numpy==1.24.1
pycparser==2.21
pyserial==3.5
python-dotenv==0.21.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
scipy==1.9.3
seeed-python-dht==0.0.2
six==1.16.0
smbus-cffi==0.5.1
smbus2==0.4.2
spidev==3.6
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
